+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - AWS ElastiCache Cluster
 - AWS ElastiCache Subnet Group
+- AWS IAM Group Membership
 
 # [v0.1.0](https://github.com/dtan4/terraforming/releases/tag/v0.1.0) (2015-06-20)